refactor: changed from using awt.Shapes from Renderers into using a rendering function. Moved SimpleRenderer into agentseek subproject
