+++ agentseek/src/main/kotlin/io/github/agentseek/util/factories/Scenes.kt
@@ -5,7 +5,7 @@ import io.github.agentseek.components.ConstantAccelerationComponent
 import io.github.agentseek.core.Scene
 import io.github.agentseek.core.engine.GameEngine
 import io.github.agentseek.util.factories.SceneFactory.emptyScene
-import io.github.agentseek.view.swing.SimpleRenderer
+import io.github.agentseek.view.gui.GameGui
 import kotlin.random.Random
 
 object Scenes {
@@ -23,7 +23,7 @@ object Scenes {
                         Vector2d(Random.nextDouble(3.0), Random.nextDouble(3.0))
                     )
                 }
-                .renderer(SimpleRenderer())
+                .renderer(GameGui.defaultRenderer())
                 .build()
             emptyScene.world.addGameObject(go)
         }
