+++ engine/src/main/kotlin/io/github/agentseek/view/View.kt
@@ -18,4 +18,10 @@ interface View {
      * Draws something on the View. Can be anything depending on its implementation.
      */
     fun draw(obj: Any)
+
+    /**
+     * Returns a default renderer
+     */
+    fun defaultRenderer(): Renderer
+
 }
